implemented the filter feature
