+++ src/components/HomePage/Category.js
@@ -1,14 +1,16 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import { faCircleArrowRight } from '@fortawesome/free-solid-svg-icons';
 import { Link } from 'react-router-dom';
 import styles from './Category.module.css';
 import { actions } from '../../redux/covid/covid';
-import IMAGES from '../../images/images';
+import data from '../../data';
 import loader from '../../images/1492.gif';
 
 function CovidCase() {
+  const [Filter, setFilter] = useState('');
+  const [Items, setItems] = useState(data);
   const dispatch = useDispatch();
   const loading = useSelector((state) => state.covidCases.loading);
   const handleCovidClick = (name) => {
@@ -23,65 +25,28 @@ function CovidCase() {
     );
   }
 
-  const data = [
-    {
-      id: 1,
-      name: 'Afghanistan',
-      image: IMAGES.Afghanistan,
-    },
-    {
-      id: 2,
-      name: 'Albania',
-      image: IMAGES.albania,
-    },
-    {
-      id: 3,
-      name: 'Algeria',
-      image: IMAGES.algeria,
-    },
-    {
-      id: 4,
-      name: 'Andorra',
-      image: IMAGES.andorra,
-    },
-    {
-      id: 5,
-      name: 'Angola',
-      image: IMAGES.angola,
-    },
-    {
-      id: 6,
-      name: 'Antarctica',
-      image: IMAGES.antarctica,
-    },
-    {
-      id: 7,
-      name: 'Antigua and Barbuda',
-      image: IMAGES.antigua,
-    },
-    {
-      id: 8,
-      name: 'Argentina',
-      image: IMAGES.argentina,
-    },
-    {
-      id: 9,
-      name: 'Armenia',
-      image: IMAGES.armenia,
-    },
-    {
-      id: 10,
-      name: 'Australia',
-      image: IMAGES.australia,
-    },
-  ];
+  const handleSubmit = (e) => {
+    e.preventDefault();
+    const input = Filter[0].toUpperCase() + Filter.slice(1);
+    const values = data.filter((item) => item.name === input);
+    setItems(values);
+    setFilter('');
+  };
+  const handleChange = (e) => {
+    setFilter(e.target.value);
+  };
+
   return (
     <div className={styles.covidCase}>
       <div className={styles.heading}>
         <h3>Stats of Cases by Countries</h3>
+        <form onSubmit={handleSubmit}>
+          <input type="text" value={Filter} onChange={handleChange} />
+          <button type="submit">Submit</button>
+        </form>
       </div>
       <div className={styles.container}>
-        {data.map((one) => (
+        {Items.map((one) => (
           <div key={one.id} className={styles.covidHome}>
             <ul className={styles.div}>
               <img src={one.image} className={styles.img} alt={one.name} />
