+++ src/data.js
@@ -1,0 +1,56 @@
+import IMAGES from './images/images';
+
+const data = [
+  {
+    id: 1,
+    name: 'Afghanistan',
+    image: IMAGES.Afghanistan,
+  },
+  {
+    id: 2,
+    name: 'Albania',
+    image: IMAGES.albania,
+  },
+  {
+    id: 3,
+    name: 'Algeria',
+    image: IMAGES.algeria,
+  },
+  {
+    id: 4,
+    name: 'Andorra',
+    image: IMAGES.andorra,
+  },
+  {
+    id: 5,
+    name: 'Angola',
+    image: IMAGES.angola,
+  },
+  {
+    id: 6,
+    name: 'Antarctica',
+    image: IMAGES.antarctica,
+  },
+  {
+    id: 7,
+    name: 'Antigua and Barbuda',
+    image: IMAGES.antigua,
+  },
+  {
+    id: 8,
+    name: 'Argentina',
+    image: IMAGES.argentina,
+  },
+  {
+    id: 9,
+    name: 'Armenia',
+    image: IMAGES.armenia,
+  },
+  {
+    id: 10,
+    name: 'Australia',
+    image: IMAGES.australia,
+  },
+];
+
+export default data;
